+++ tests/unit/domestic/test_templatetags.py
@@ -1,10 +1,13 @@
 from unittest import mock
 
 import pytest
+from bs4 import BeautifulSoup
+from django.core.paginator import Paginator
 from django.template import Context, Template
 
 from domestic.templatetags.component_tags import (
     get_meta_description,
+    get_pagination_url,
     industry_accordion_case_study_is_viable,
     industry_accordion_is_viable,
     parse_date,
@@ -325,3 +328,84 @@ def test_get_meta_description(attrs_to_set, expected):
         setattr(page, attr, value)
 
     assert get_meta_description(page) == expected
+
+
+@pytest.mark.parametrize(
+    'url,expected',
+    (
+        ('/foo/bar/', '/foo/bar/?'),
+        ('/foo/bar/?page=2', '/foo/bar/?'),
+        ('/foo/bar/?page=2&baz=3', '/foo/bar/?baz=3&'),
+        ('/foo/bar/?bam=test&page=2&baz=3', '/foo/bar/?bam=test&baz=3&'),
+    ),
+)
+def test_get_pagination_url(rf, url, expected):
+    request = rf.get(url)
+    assert get_pagination_url(request, 'page') == expected
+
+
+@pytest.mark.parametrize(
+    'count,current,expected',
+    (
+        (21, 1, '[1] 2 3 4 5 N'),
+        (21, 2, 'P 1 [2] 3 4 5 N'),
+        (21, 3, 'P 1 2 [3] 4 5 N'),
+        (21, 4, 'P 1 2 3 [4] 5 N'),
+        (21, 5, 'P 1 2 3 4 [5]'),
+        (30, 1, '[1] 2 3 4 5 6 N'),
+        (40, 1, '[1] 2 3 4 ... 8 N'),
+        (40, 2, 'P 1 [2] 3 4 ... 8 N'),
+        (40, 3, 'P 1 2 [3] 4 ... 8 N'),
+        (40, 4, 'P 1 2 3 [4] ... 8 N'),
+        (40, 5, 'P 1 ... [5] 6 7 8 N'),
+        (40, 6, 'P 1 ... 5 [6] 7 8 N'),
+        (40, 7, 'P 1 ... 5 6 [7] 8 N'),
+        (40, 8, 'P 1 ... 5 6 7 [8]'),
+        (60, 1, '[1] 2 3 4 ... 12 N'),
+        (60, 2, 'P 1 [2] 3 4 ... 12 N'),
+        (60, 3, 'P 1 2 [3] 4 ... 12 N'),
+        (60, 4, 'P 1 2 3 [4] ... 12 N'),
+        (60, 5, 'P 1 ... 4 [5] 6 ... 12 N'),
+        (60, 6, 'P 1 ... 5 [6] 7 ... 12 N'),
+        (60, 7, 'P 1 ... 6 [7] 8 ... 12 N'),
+        (60, 8, 'P 1 ... 7 [8] 9 ... 12 N'),
+        (60, 9, 'P 1 ... [9] 10 11 12 N'),
+        (60, 10, 'P 1 ... 9 [10] 11 12 N'),
+        (60, 11, 'P 1 ... 9 10 [11] 12 N'),
+        (60, 12, 'P 1 ... 9 10 11 [12]'),
+    ),
+)
+def test_pagination(count, current, expected, rf):
+    template = Template(
+        """
+            {% load pagination from component_tags %}
+            {% pagination pagination_page=pagination_page %}
+        """
+    )
+
+    page_size = 5
+    objects = [item for item in range(count)]
+
+    paginator = Paginator(objects, page_size)
+    pagination_page = paginator.page(current)
+    context = {'request': rf.get('/'), 'pagination_page': pagination_page}
+
+    html = template.render(Context(context))
+
+    soup = BeautifulSoup(html, 'html.parser')
+
+    items = []
+    if soup.findAll('a', {'class': 'pagination-previous'}):
+        items.append('P')
+    for element in soup.find_all('li'):
+        if element.find('span'):
+            items.append('...')
+        else:
+            button = element.find('a')
+            if 'button' in button['class']:
+                items.append(f'[{button.string}]')
+            else:
+                items.append(button.string)
+    if soup.findAll('a', {'class': 'pagination-next'}):
+        items.append('N')
+    assert ' '.join(items) == expected